backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
anthropic
python-dotenv
python-multipart
